#79 Augment Window is close to creating augment prefabs
Haven't pushed in a while. Sorry.
This push contains a lot of structs and random information for Augments
There's an augment parser which will take JSON data for augments and make it into actual data to be used in structs.
There's an augment Cretor which will (eventually) actually create the augments from data parsed in. (Haha. Programming jokes. Kill me.)
AugmentGo is a game object representation of an Augment, with a mesh and material.